assert and skip writing cache
When we json_encode to send to the cache, it might fail. Usually happens when there is a faulty UTF-8 conversion.
We can catch this by looking at the result of json_encode. We can keep the application working, by just skipping caching write. And we notify the developer by asserting.